new 2018 tooling
remove grunt and build in preference for simply delivering src
move to jest and eslint
