+++ .gitignore
@@ -15,3 +15,4 @@ results
 npm-debug.log
 node_modules
 secrets
+.vscode
