+++ .travis.yml
@@ -1,6 +1,7 @@
 sudo: false
 language: node_js
-node_js: 
-  - "0.12"
-  - "0.10"
-before_install: npm install -g grunt-cli
+node_js:
+  - "9.2.1"
+script:
+  - yarn run lint
+  - yarn run test
