--- tests/tests.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>objectDiff</title>
-  <link rel="stylesheet" href="vendor/qunit.css">
-</head>
-<body>
-  <div id="qunit"></div>
-  <div id="qunit-fixture"></div>
-  <script src="vendor/jquery.js" type="text/javascript"></script>
-  <script src="vendor/qunit.js"></script>
-  <script src="../dist/unmin/js/objectDiff.js" type="text/javascript"></script>
-  <script src="unit/objectDiff.js" type="text/javascript"></script>
-</body>
-</html>
